Update src/main/java/io/github/ardoco/textproviderjson/converter/DtoToObjectConverter.java
1L instead of (long) 1

+++ src/main/java/io/github/ardoco/textproviderjson/converter/DtoToObjectConverter.java
@@ -103,7 +103,7 @@ public class DtoToObjectConverter {
     }
 
     private DependencyImpl convertIncomingDependency(IncomingDependencyDTO dependencyDTO) {
-        return new DependencyImpl(dependencyDTO.getDependencyTag(), dependencyDTO.getSourceWordId() - (long) 1);
+        return new DependencyImpl(dependencyDTO.getDependencyTag(), dependencyDTO.getSourceWordId() - 1L);
     }
 
     private DependencyImpl convertOutgoingDependency(OutgoingDependencyDTO dependencyDTO) {
